yantra-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
chromadb
openai
